split offline related to a separate file
