+++ flow-client/src/main/resources/META-INF/resources/frontend/Connect.ts
@@ -1,5 +1,5 @@
 /* tslint:disable:max-classes-per-file */
-import { DBSchema, IDBPDatabase, openDB } from 'idb';
+import { DeferrableResult, DeferredCallHandler, EndpointRequest, offline } from './Offline';
 
 const $wnd = window as any;
 $wnd.Vaadin = $wnd.Vaadin || {};
@@ -7,9 +7,6 @@ $wnd.Vaadin.registrations = $wnd.Vaadin.registrations || [];
 $wnd.Vaadin.registrations.push({
   is: 'endpoint'
 });
-
-const REQUEST_QUEUE_DB_NAME = 'request-queue';
-const REQUEST_QUEUE_STORE_NAME = 'requests';
 
 interface ConnectExceptionData {
   message: string;
@@ -163,15 +160,6 @@ export class ValidationErrorData {
 }
 
 /**
- * The callback for deferred calls
- */
-export type OnDeferredCallCallback = (deferrableCall: DeferredCall) => Promise<void>;
-
-export interface DeferredCallHandler {
-  handleDeferredCall: OnDeferredCallCallback;
-}
-
-/**
  * The `ConnectClient` constructor options.
  */
 export interface ConnectClientOptions {
@@ -186,9 +174,9 @@ export interface ConnectClientOptions {
   middlewares?: Middleware[];
 
   /**
-   * The `onDeferredCall` property value
+   * The `deferredCallHandler` property valueThe
    */
-  onDeferredCall?: OnDeferredCallCallback
+  deferredCallHandler?: DeferredCallHandler;
 }
 
 /**
@@ -289,10 +277,8 @@ export class ConnectClient {
   middlewares: Middleware[] = [];
 
   /**
-   * The callback for deferred calls
+   * The handler for handling deferred calls
    */
-  onDeferredCall?: OnDeferredCallCallback;
-
   deferredCallHandler?: DeferredCallHandler;
 
   /**
@@ -307,7 +293,7 @@ export class ConnectClient {
       this.middlewares = options.middlewares;
     }
 
-    this.onDeferredCall = options.onDeferredCall;
+    this.deferredCallHandler = options.deferredCallHandler;
 
     this.processDeferredCalls = this.processDeferredCalls.bind(this);
 
@@ -344,31 +330,18 @@ export class ConnectClient {
     method: string,
     params?: any,
   ): Promise<DeferrableResult<any>> {
-    if (this.checkOnline()) {
+    if (offline.checkOnline()) {
       const result = await this.call(endpoint, method, params);
       return { isDeferred: false, result };
     } else {
       let endpointRequest:EndpointRequest = { endpoint, method, params };
-      endpointRequest = await this.cacheEndpointRequest(endpointRequest);
+      endpointRequest = await offline.cacheEndpointRequest(endpointRequest);
       return { isDeferred: true, endpointRequest };
     }
   }
 
   async processDeferredCalls() {
-    const db = await this.openOrCreateDB();
-
-    /**
-     * Cannot wait for submitting the cached requests in the indexed db transaction,
-     * as the transaction only wait for db operations.
-     * See https://github.com/jakearchibald/idb#transaction-lifetime
-     */
-    const shouldSubmit = await this.shouldSubmitCachedRequests(db);
-
-    if (shouldSubmit) {
-      await this.submitCachedRequests(db);
-    }
-
-    db.close();
+    await offline.processDeferredCalls(this.requestCall.bind(this), this.deferredCallHandler);
   }
 
   private async requestCall(
@@ -462,78 +435,6 @@ export class ConnectClient {
     return chain(initialContext);
   }
 
-  private checkOnline(): boolean {
-    return navigator.onLine;
-  }
-
-  private async cacheEndpointRequest(endpointRequest: EndpointRequest): Promise<EndpointRequest>{
-    const db = await this.openOrCreateDB();
-    const id = await db.add(REQUEST_QUEUE_STORE_NAME, endpointRequest);
-    db.close();
-    endpointRequest.id = id;
-    return endpointRequest;
-  }
-
-  private async openOrCreateDB(): Promise<IDBPDatabase<RequestQueueDB>> {
-    return openDB<RequestQueueDB>(REQUEST_QUEUE_DB_NAME, 1, {
-      upgrade(db) {
-        db.createObjectStore(REQUEST_QUEUE_STORE_NAME, {
-          keyPath: 'id',
-          autoIncrement: true
-        });
-      },
-    });
-  }
-
-  private async shouldSubmitCachedRequests(db: IDBPDatabase<RequestQueueDB>) {
-    let shouldSubmit = false;
-    if (db.objectStoreNames.contains(REQUEST_QUEUE_STORE_NAME) && await db.count(REQUEST_QUEUE_STORE_NAME) > 0) {
-      const tx = db.transaction(REQUEST_QUEUE_STORE_NAME, 'readwrite');
-
-      let cursor = await tx.store.openCursor();
-      while (cursor) {
-        const request = cursor.value;
-        if (!request.submitting) {
-          shouldSubmit = true;
-          request.submitting = true;
-          cursor.update(request);
-        }
-        cursor = await cursor.continue();
-      }
-      await tx.done;
-    }
-    return shouldSubmit;
-  }
-
-  
-  private async submitCachedRequests(db: IDBPDatabase<RequestQueueDB>) {
-    const cachedRequests = await db.getAll(REQUEST_QUEUE_STORE_NAME);
-    for (const request of cachedRequests) {
-      if (request.submitting) {
-        try {
-          let shouldDelete = true;
-          if (this.deferredCallHandler) {
-            const deferredCall = new DeferredCall(request, ({endpoint, method, params}) => this.requestCall.bind(this)(true, endpoint, method, params));
-            await this.deferredCallHandler.handleDeferredCall(deferredCall);
-            shouldDelete = !deferredCall._shouldKeepInTheQueue();
-          } else {
-            await this.requestCall(true, request.endpoint, request.method, request.params);
-          }
-          if(shouldDelete){
-            await db.delete(REQUEST_QUEUE_STORE_NAME, request.id!);
-          }else{
-            request.submitting = false;
-            await db.put(REQUEST_QUEUE_STORE_NAME, request);
-          }
-        } catch (error) {
-          request.submitting = false;
-          await db.put(REQUEST_QUEUE_STORE_NAME, request);
-          throw error;
-        }
-      }
-    }
-  }
-
   // Re-use flow loading indicator when fetching endpoints
   private loading(action: boolean) {
     if ($wnd.Vaadin.Flow?.loading) {
@@ -542,50 +443,3 @@ export class ConnectClient {
   }
 }
 
-export class DeferredCall implements EndpointRequest{
-  id?: number;
-  endpoint: string;
-  method: string;
-  params?: any;
-  submitting?: boolean;
-  private _keepInTheQueue = false;
-  private _submitFunction: (call: EndpointRequest)=>Promise<any>;
-  constructor(endpointCall: EndpointRequest, submitFunction: (call: EndpointRequest)=>Promise<any>){
-    this.id = endpointCall.id;
-    this.endpoint = endpointCall.endpoint;
-    this.method = endpointCall.method;
-    this.params = endpointCall.params;
-    this._submitFunction = submitFunction;
-  }
-  
-  async submit(): Promise<any>{
-    return this._submitFunction(this);
-  }
-
-  keepInTheQueue(){
-    this._keepInTheQueue = true;
-  }
-  _shouldKeepInTheQueue(){
-    return this._keepInTheQueue;
-  }
-}
-export interface EndpointRequest {
-  id?: number;
-  endpoint: string;
-  method: string;
-  params?: any;
-  submitting?: boolean
-}
-
-export interface DeferrableResult<T> {
-  isDeferred: boolean;
-  endpointRequest?: EndpointRequest;
-  result?: T;
-}
-
-interface RequestQueueDB extends DBSchema {
-  requests: {
-    value: EndpointRequest;
-    key: number;
-  };
-}
